# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: loan-application
  description: My cool app
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
definitions:
  ApplicationBody:
    type: object
    title: application
    required:
      - applicant_first_name
      - applicant_last_name
      - loan_amount
    properties:
      applicant_first_name:
        type: string
        example: Martin
      applicant_last_name:
        type: string
        example: Flower
      loan_amount:
        type: number
        example: 100000
paths:
...
  /applications/{id}:
    get:
      tags:
        - Applications
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the application to retrieve
          type: string
      responses:
        200:
          description: Return the application with the specified id
        404:
          description: Application not found
    delete:
      tags:
        - Applications
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the application to delete
          type: string
      responses:
        204:
          description: Return the status of the request
        404:
          description: Application not found
    put:
      tags:
        - Applications
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the application to update
          type: string
        - name: application
          in: body
          description: an updated application
          required: true
          schema: 
            $ref: "#/definitions/ApplicationBody"
      responses:
        200:
          description: Return the application with the specified id
        404:
          description: Application not found
...